requirements.txt
mistralai>=1.0.0
pypandoc>=1.12
natsort>=8.4.0
pyyaml>=6.0
PyQt6>=6.6.0
python-dotenv>=1.0.0
